# shift-left-demo/spring-boot-order-service-devtools/src/main/resources/order-service-openapi.yaml
---
openapi: 3.0.2
info:
  title: Order Service API
  version: 0.1.0
  description: Simple Order Service API
  contact:
    name: Laurent Broudoux
    url: https://github.com/lbroudoux
    email: laurent@microcks.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
paths:
  /orders:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInfo'
            examples:
              valid_order:
                value:
                  customerId: lbroudoux
                  productQuantities:
                  - productName: Millefeuille
                    quantity: 1
                  - productName: Eclair Cafe
                    quantity: 2
                  totalPrice: 9.4
              invalid_order:
                value:
                  customerId: lbroudoux
                  productQuantities:
                  - productName: Millefeuille
                    quantity: 1
                  - productName: Eclair Chocolat
                    quantity: 1
                  totalPrice: 4.8
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInfo'
              examples:
                valid_order:
                  value:
                    id: 5455c8e8-087a-426e-8440-65c8c005d871
                    status: CREATED
                    customerId: lbroudoux
                    productQuantities:
                    - productName: Millefeuille
                      quantity: 1
                    - productName: Eclair Cafe
                      quantity: 2
                    totalPrice: 9.4
          description: Order is correct and has been created
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnavailableProduct'
              examples:
                invalid_order:
                  value:
                    productName: Eclair Chocolat
                    details: Eclair Chocolat are not available at the moment
          description: "Order cannot be processed because of a validation error (ex:\
            \ unavailable product)"
      operationId: PlaceOrder
      summary: Place a new Order
      description: Place a new Order in the system. Will perform extra checks before
        saving Order to detect invalid demand
components:
  schemas:
    OrderInfo:
      description: Represents info needed for creating an Order
      required:
      - customerId
      - productQuantities
      - totalPrice
      type: object
      properties:
        customerId:
          description: Identifier of customer of this order
          type: string
        productQuantities:
          description: Desired products and quantities for this order
          type: array
          items:
            $ref: '#/components/schemas/ProductQuantity'
        totalPrice:
          format: double
          description: Total price of the order
          type: number
    ProductQuantity:
      description: Association of product name with quantity
      required:
      - productName
      - quantity
      type: object
      properties:
        productName:
          description: Desired product name
          type: string
        quantity:
          description: Desired quantity
          type: integer
    Order:
      description: Full created Order with all informations
      type: object
      allOf:
      - required:
        - id
        - status
        type: object
        properties:
          id:
            description: Unique identifier of order
            type: string
          status:
            description: Status of Order
            enum:
            - CREATED
            - VALIDATED
            - CANCELED
            - FAILED
            type: string
      - $ref: '#/components/schemas/OrderInfo'
    UnavailableProduct:
      description: ""
      required:
      - productName
      type: object
      properties:
        productName:
          description: ""
          type: string
        details:
          description: Details of unavailability
          type: string